Modal to add cat information has been added.
The modal is basically a <div> that has been created with Bootstrap's
help for the css.

The modal that was previously available has now been modified.
It now offers the fields to add cat information, which are: 4 image
inputs and one text input for the name.

It is possible to add more fields in the future, this requires to
manipulate django views and forms. For further information about how to
proceed in this case please refer to the django documentation, with
@PirLucas who has implemented a big part of the views and forms, or with me.

The next step is to manage the django views and forms to add the cat.

Some javascript is going to be implemented to add the cat to the main
profile page without reloading it.