The issue #20 is in progress. Files added.
Following files are added:
1. showCatImg.js
2. cat_PNG50482-2467262089.png <-- Explained below (*)
3. cat_PNG50482-2467262089_zpbpexA.png <-- Explained below (*)

Following files are modified:
1. Code/settings.py
2. CICO/urls.py
3. CICO/views.py
4. CICO/models.py
5. CICO/forms.py
6. templates/CICO/profileIndex.html

added showCatImg.js file.
This js file has a function that lets the user to see the image of the cat they are about to submit.
This function is called when the user uploads the image of the cat.

* Media file and cat pictures

The cat images are the beginning of some testing. After some questions about
how the cat images should be stored in the database. Instead of adding the images
to the database, I decided to add the images to the media folder. The images are
stored in the media folder and the path to the image SHOULD BE stored in the database.
This is a start of the logic that will be tested. This way we will prevent
the overloading of the database with images, this will also make the database
smaller and faster.

What is trying to be accomplished is that the user will upload the image of the cat
and the image will be stored in the media folder. The path to the image will be stored
in the database. When the user wants to see the image of the cat, the path will be
retrieved from the database and the image will be displayed.

We have to be sure that the other users cat information are not available to other users.
This is why further storing logic of the folders is in progress.

The logic is thought as following:
- A folder shall be created for each user. Inside this initial folder
    there will be a folder for each cat. The cat folder will contain the image of the cat.
    The path to the image will be stored in the database.
    The name of the cat folder will be created in base of cat ID.
    Still thinking on how to automatically create a folder name based on the
    user ID and the cat ID.
- The user will be able to see the image of the cat by retrieving the path from the database.
    The path will be retrieved based on the cat ID.

Other files modified:

- Code/settings.py
    - Added the following lines:
        MEDIA_ROOT = os.path.join(BASE_DIR, 'media')
        MEDIA_URL = '/media/'
    What this is going to do is to create a media folder in the root of the project.
    This folder will be used to store the images of the cats.

- CICO/urls.py
    - Added the following lines:
        from django.conf import settings
        from django.conf.urls.static import static
        urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
    This is going to allow the media files to be served by the development server.

    And also the line:
        path('add_cat/', views.def_cat, name='add_cat'),
    This is going to allow the user to add a cat.

- CICO/views.py
    - Added the following lines:
        from django.conf import settings
        from django.core.files.storage import FileSystemStorage
    This is going to allow the media files to be served by the development server.

    And the def_cat function it's working so that if the form is valid adn the user is logged in, the image
    will be stored in the media folder. This returns a json response which
    is going to tell us if the image was stored or not.

- CICO/models.py
    Added the     image = models.ImageField(upload_to='cats/') line. Inside
    of the Cat class. This is going to allow the image to be stored in the
    media folder.

- CICO/forms.py
    Added the Class catForm(forms.ModelForm): line. That should get the image
    and the name of the cat from the form that's in the HTML.

- templates/CICO/profileIndex.html
    Added some scripting for testing the image upload. This is going to be
    removed later on. Or modified to be used for the final version.