Changing the way the cat ID is generated #20.
Now the cat ID is generated using UUID4.
This way we can avoid the problem of having the same ID for two different cats.

The cat's information is being saved correctly into the database.

The way the cat's image is being saved into the media folder is not correct yet.
The system I'm planning to use is to create a folder for each user automatically,
after that, when he uploads a cat's image, this action should create a folder for the cat
inside the user's folder and save the image there.

The next step is to display the cat's information in the profile page.