Showing cats feature in progress - #20
It is still in progress, it is not working yet.

- loadCats.js:
It is a script that loads the cats from the database and displays them in the page.
It is called by the profileIndex.html template.

- serializers.py:
It is a serializer that serializes the cats from the database.

- views.py:
Added class UserCatList that returns the cats of the user.
The function get_queryset() returns the cats of the user.

Added the function get_cats_for_user that returns the cats of the user.
It takes in parameter a request and returns a list of cats.
It has the decorator @login_required to make sure that the user is logged in.

- urls.py:
Added the url for the function get_cats_for_user. It is still not working,
this has to be correctly implemented.

- templates/CICO/profileIndex.html:
Added the script loadCats.js to the template. And also a hidden input to
store the user id. Still thinking if it is the best way to do it.

- templates/baseProfil.html:
Added the script loadCats.js to the template.