Getting the cat Id for the Post request
To show in the post request the catId I added in the function add_cat
the line response.catId, so when the post request is done the catId is
shown in the response. This interacts with the AJAX function in the
profileIndex.html file.

The AJAX function is called when the user clicks the button to add a
cat. The function gets the catId from the response and then it calls
the function add_cat_details with the catId as parameter.