Load cat details in dev_Carlos (#80)
* Modifying format

* formatting page

* Getting the cat Id for the Post request

To show in the post request the catId I added in the function add_cat
the line response.catId, so when the post request is done the catId is
shown in the response. This interacts with the AJAX function in the
profileIndex.html file.

The AJAX function is called when the user clicks the button to add a
cat. The function gets the catId from the response and then it calls
the function add_cat_details with the catId as parameter.

* Added response.catId line in profileIndex.html

* Loading cat details in modal

In urls.py the url for the modal is defined as:
    path('get_cat_details/<uuid:catId>/', views.get_cat_details, name='get_cat_details'),

We are using the UUID of to cat to get the info from the server.

In views.py created new function to get the cat details:
    def get_cat_details(request, catId):
    # Fetch the cat details
    cat = Cats.objects.get(ownerId_id=request.user, catId=catId)
    return JsonResponse({
        'name': cat.name,
        'image_url': cat.image.url if cat.image else ''
    })

When the cat button is clicked a request is done, we fetch the info of the cats
for the current user and we fetch the info of the cat that corresponds with this
ID. it returs the name and the image of the cat.

In profileIndex.html :
Added a modal to load the data of the cat (name and image) inside of it.
The modal is hidden by default and it is shown when the button is clicked.
The button has the id of the cat so we can fetch the info of the cat.

Changed the way that the function appendCatRow works, it works with objects  as parameters now.

Added a script called with the function fetchAndDisplayCatDetails that fetches the info of the cat
and displays it in the modal.
It works with the url and having the catId as parameter to query into the database and get the info of the cat.