US38 - Issue #23 - Cat deletion feature done
The Cat deletion feature has been done, there are other small modifications
done for other features, but they are not important. I will quickly talk
about that.

In views.py:
- The cat deletion feature is called delete_cat, it gets the request and
catId. I created a variable called cat that calls the function get_object_or_404
if the cat exists it returns the object of the cat, if not returns error 404.

If there is the image of the cat in the database, it will delete it from the
from the folder and then from the database. Then it will delete the cat from the database.

The message 'cat deleted' will be shown if everything goes as expected.

in urls.py:
- I added the url for the delete_cat feature. It takes the catId as string and
not as UUID. The catId is saved without hyphens in the DB. There was an error concerning this.
This is why it takes the UUID as string and not as UUID.

EXTRA: I modified the get_cat_details feature to work the same way.

in profileIndex.html:
- Added ajax request to delete the cat. It takes the catId as string. When you open the cat modal the UUID
is added to the modal as data-cat-id. When you click on the delete button, it will get the catId from the
modal and send it to the delete_cat feature.

EXTRA: changed the way the handleAddCatError function works. It gets the response from the server instead
of a xhr function.

Added some tests for the delete_cat feature and the add cat feature.