Fix the modify function.
Modify function is already working, the URL has been added in urls.py.

The views now contain the function to add the cats.

And the profileIndex.html has been modified, modal with the fields to modify the cat has been added.
And the ajax requests for this function has been added.